+++ cuBERT_topic_modelling/setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="cubertopic",
+    version='1.0',
+    packages=find_packages()
+)
