+++ cuBERT_topic_modelling/tests/test_data_preprocess.py
@@ -1,15 +1,7 @@
 import pandas as pd
-import sys
-from os.path import dirname, realpath
 import cudf
 import re
-
-filepath = realpath(__file__)
-dir_of_file = dirname(filepath)
-parent_dir_of_file = dirname(dir_of_file)
-parents_parent_dir_of_file = dirname(parent_dir_of_file)
-sys.path.append(parents_parent_dir_of_file + "/cuBERT-topic-modelling/")
-from vectorizer import CountVecWrapper
+from vectorizer.vectorizer import CountVecWrapper
 
 
 data = [
