+++ cuBERT_topic_modelling/tests/test_hdbscan_clustering.py
@@ -1,15 +1,7 @@
-import sys
-from os.path import dirname, realpath
 import cudf
 from sklearn.datasets import fetch_20newsgroups, make_blobs
 import pytest
-
-filepath = realpath(__file__)
-dir_of_file = dirname(filepath)
-parent_dir_of_file = dirname(dir_of_file)
-parents_parent_dir_of_file = dirname(parent_dir_of_file)
-sys.path.append(parents_parent_dir_of_file + "/cuBERT-topic-modelling/")
-from cuBERTopic import gpu_bertopic
+from cuBERTopic import gpu_BERTopic
 
 
 newsgroup_docs = fetch_20newsgroups(
@@ -43,7 +35,7 @@ def test_hdbscan_cluster_embeddings(samples, features, centers):
     old_df = cudf.DataFrame(
         {"Document": documents, "ID": range(len(documents)), "Topic": None}
     )
-    gpu_topic = gpu_bertopic()
+    gpu_topic = gpu_BERTopic()
     new_df, _ = gpu_topic.clustering_hdbscan(embeddings, old_df)
 
     assert len(new_df.Topic.unique()) == centers
