+++ cuBERT_topic_modelling/tests/test_umap_dr.py
@@ -1,15 +1,7 @@
-import sys
-from os.path import dirname, realpath
 import numpy as np
 from sklearn.datasets import fetch_20newsgroups
 import pytest
-
-filepath = realpath(__file__)
-dir_of_file = dirname(filepath)
-parent_dir_of_file = dirname(dir_of_file)
-parents_parent_dir_of_file = dirname(parent_dir_of_file)
-sys.path.append(parents_parent_dir_of_file + "/cuBERT-topic-modelling/")
-from cuBERTopic import gpu_bertopic
+from cuBERTopic import gpu_BERTopic
 
 newsgroup_docs = fetch_20newsgroups(
     subset="all", remove=("headers", "footers", "quotes")
@@ -24,7 +16,7 @@ def test_umap_reduce_dimensionality(embeddings, shape):
     as the main goal here is to reduce the dimensionality, the quality is
     tested in the full pipeline.
     """
-    gpu_topic = gpu_bertopic()
+    gpu_topic = gpu_BERTopic()
     umap_embeddings = gpu_topic.reduce_dimensionality(embeddings)
     assert umap_embeddings.shape == (shape, 5)
